+++ CHANGELOG.md
@@ -14,6 +14,9 @@ Remember that the keywords that you can use are the following:
 
 ## Unreleased
 
+### bugfix
+- K8s supervisor: force a Flux reconciliation on stalled HelmReleases (e.g. after exhausting install/upgrade retries) when a new remote config
+
 ## v1.22.0 - 2026-08-18
 
 ### 🚀 Enhancements